+++ crates/web-client/test/playwright.global.setup.ts
@@ -24,6 +24,7 @@ export const test = base.extend<{ forEachTest: void }>({
       });
 
       await page.goto("http://localhost:8080");
+
       await page.evaluate(
         async ({ MIDEN_NODE_PORT, remoteProverPort }) => {
           // Import the sdk classes and attach them
@@ -32,6 +33,7 @@ export const test = base.extend<{ forEachTest: void }>({
           for (const [key, value] of Object.entries(sdkExports)) {
             window[key] = value;
           }
+
           let rpcUrl = `http://localhost:${MIDEN_NODE_PORT}`;
           let proverUrl = remoteProverPort
             ? `http://localhost:${remoteProverPort}`